Make production & dev builds succeed after upgrading webpack
The `colors` argument to webpack needed to be `color` for `yarn run dev`
to succeed.

The linter highlighted `postMessage` as an already defined built-in
global variable.

The UglifyJsPlugin was removed from webpack 4 so that had to be added back
in separately as a dev dependency and configured in a slightly different
way as per the new webpack version.