Add types for GDALDataset
The `Args` type is currently very permissive to support the args that
might be passed into `render`.

Most of the annotations here are straightforward. I removed the unused
`reject` callbacks from the promises since my linter yelled at me about
that.

I did make a small change to `cliOrderArgs` to use the spread operator
instead of `concat` only because the type checker was getting confused
by `concat` for some reason.