Fix tests
by making the entrypoint be `index.js` instead of `index.ts`. Before
this change, tests were failing with:

TypeError: loam.initialize is not a function

Karma seems to have trouble with the TypeScript file since it doesn't
have a TS -> JS build step as part of running tests. For expediency's
sake, let's just leave it as a JS file for now since all it does is
import some functions/classes and export them.