Validate defaultPrefix is a valid URL
If it isn't, use undefined instead. An invalid (but non-undefined) URL
will always cause loam.initialize() to throw an exception, even if the
user attempts to provide a valid URL as a parameter to initialize().
This now validates defaultPrefix before attempting to use it, and
uses undefined if the defaultPrefix is not a valid URL. The user in this
case will be _required_ to provide a valid absolute URL parameter to
initialize(), because there is no default to fall back on.